initramfs: work around initramfs-tools bug #688794
Use "_" in place of "+-." for the initramfs script name.  This works
around a PANIC during boot when the initramfs was created in a system
with noexec $TMPDIR.

initramfs-tools installs code to the initramfs that is highly alergic to
script names that have characters not acceptable on shell variable
names.

While at it, drop useless initramfs dependency on udev, as we're running
in init-premount, which implies udev has already run.