load microcode module on package install/upgrade
When the package is first installed, manually loading the microcode
module might be necessary to be able to update the microcode without a
reboot.

Should we succeed to load the microcode module, there is no need to
manually trigger a microcode update, as the module load will trigger it
by itself.

Also add a script to make sure the microcode module is loaded before
kernel updates, otherwise we will require a reboot before we can update
the microcode once again...