blacklist automated loading of the microcode module
We only want the microcode to be automatically updated by the initramfs,
which is known to not trigger kernel bugs in the amd microcode update
kernel drivers.

This is in line with the desired behavior of only updating microcode
*automatically* during system boot, when it is safer to do so.  The
local admin can still load the microcode module and update the microcode
manually at any time, of course.

This is in sync with the intel-microcode packages in Debian jessie,
which will also blacklist the microcode module.