initramfs.hook: don't force_load microcode.ko when missing
Detect a missing microcode.ko and don't attempt to force_load() it,
otherwise we get spurious warnings at boot.  In verbose mode, log
the fact that the microcode driver is modular.

For Linux 4.4 and later, skip the entire module loading logic, since the
microcode driver cannot be modular for those kernels.

Closes: #809444